test_get_withAddress, test_update, test_delete_notPresent ( ok ) test_insert_withAddress ( il test e' verde ma non aggiunge l'indirizzo nel db ) test_delete_WithAddress ( non ho avuto tempo di approfondirlo )

+++ src/main/java/org/sg/dao/impl/StudentDaoImpl.java
@@ -26,10 +26,13 @@ public class StudentDaoImpl extends GenericDao implements StudentDao {
 	}
 
 	public boolean delete(Integer id) {
-		entityManager.getTransaction().begin();
 		StudentEntity studentEntity = entityManager.find(StudentEntity.class, id);
-		entityManager.remove(studentEntity);
-		entityManager.getTransaction().commit();
-		return true;
+		if (studentEntity != null) {
+			entityManager.getTransaction().begin();
+			entityManager.remove(studentEntity);
+			entityManager.getTransaction().commit();
+			return true;
+		}
+		return false;
 	}
 }
